removed some politics
There's been recent complaints about the difficulty of getting patches into core. And it's true, I think if folks have better things, put them up and we'll use them. If enough people use them, that'll give the core team more inclination to add that sort of thing to core. But if not, and if it's good, people will use it. Clojure is open enough that you don't need your changes to go into core for us to use them. But I'll leave that politicking out of the readme...